backend/requirements.txt
aiofiles==22.1.0
anyio==3.6.2
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
fastapi==0.87.0
h11==0.14.0
httptools==0.5.0
idna==3.4
oauthlib==3.2.2
pydantic==1.10.2
python-dotenv==0.21.0
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
sniffio==1.3.0
starlette==0.21.0
tweepy==4.12.1
typing_extensions==4.4.0
urllib3==1.26.12
uvicorn==0.19.0
watchfiles==0.18.1
websockets==10.4
textblob==0.17.1
nltk==3.7
snscrape==0.4.3.20220106
pydantic==1.10.2
pymongo==4.3.3
dnspython==2.2.1
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.5
bcrypt==4.0.1
pandas==1.5.2
